src/requirements.txt
fastapi
uvicorn[standard]
aiofiles
python-multipart
numpy
pillow
opencv-python-headless
websockets
soundfile
torch
transformers
diffusers
accelerate
ftfy
sentencepiece
whisper
ffmpeg-python
tqdm
scipy
sounddevice